requirements.txt
Django==2.1.4
djangorestframework==3.9.0
